+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/enums/generic/DirectionYZ.kt
@@ -1,7 +1,7 @@
-package com.windea.breezeframework.core.enums.normal
+package com.windea.breezeframework.core.enums.generic
 
 import com.windea.breezeframework.core.annotations.messages.*
-import com.windea.breezeframework.core.enums.normal.Dimension.*
+import com.windea.breezeframework.core.enums.generic.Dimension.*
 
 /**基于YZ轴的二维方向。*/
 @Name("基于YZ轴的二维方向")