+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/AnyExtensions.kt
@@ -3,6 +3,7 @@
 package com.windea.breezeframework.core.extensions
 
 /**智能地判断两个对象是否相等。特殊对待数组类型，默认递归执行操作。*/
+@JvmOverloads
 fun Any?.smartEquals(other: Any?, deepOperation: Boolean = true) = when {
 	this !is Array<*> || other !is Array<*> -> this == other
 	!deepOperation -> this contentEquals other
@@ -10,6 +11,7 @@ fun Any?.smartEquals(other: Any?, deepOperation: Boolean = true) = when {
 }
 
 /**智能地得到当前对象的哈希码。特殊对待数组类型，默认递归执行操作。*/
+@JvmOverloads
 fun Any?.smartHashCode(deepOperation: Boolean = true) = when {
 	this !is Array<*> -> this.hashCode()
 	!deepOperation -> this.contentHashCode()
@@ -17,6 +19,7 @@ fun Any?.smartHashCode(deepOperation: Boolean = true) = when {
 }
 
 /**智能地将当前对象转化为字符串。特殊对待数组类型，默认递归执行操作。*/
+@JvmOverloads
 fun Any?.smartToString(deepOperation: Boolean = true) = when {
 	this !is Array<*> -> this.toString()
 	!deepOperation -> this.contentToString()