+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/ArrayExtensions.kt
@@ -53,15 +53,19 @@ inline fun CharArray.onEach(action: (Char) -> Unit): CharArray {
 
 
 /**将数组转化为流对象。可指定索引范围，默认为整个数组。*/
+@JvmSynthetic
 fun <T> Array<out T>.stream(start: Int = 0, end: Int = this.size): Stream<T> = Arrays.stream(this, start, end)
 
 /**将数组转化为流对象。可指定索引范围，默认为整个数组。*/
+@JvmSynthetic
 fun IntArray.stream(start: Int = 0, end: Int = this.size): IntStream = Arrays.stream(this, start, end)
 
 /**将数组转化为流对象。可指定索引范围，默认为整个数组。*/
+@JvmSynthetic
 fun LongArray.stream(start: Int = 0, end: Int = this.size): LongStream = Arrays.stream(this, start, end)
 
 /**将数组转化为流对象。可指定索引范围，默认为整个数组。*/
+@JvmSynthetic
 fun DoubleArray.stream(start: Int = 0, end: Int = this.size): DoubleStream = Arrays.stream(this, start, end)