+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/CoreExtensions.kt
@@ -9,14 +9,17 @@ import kotlin.contracts.*
 //region Standard.kt extensions (Todo functions)
 /**表明一个操作推迟了实现。*/
 @TodoMarker
+@JvmSynthetic
 inline fun DELAY() = DELAY { Unit }
 
 /**表明一个方法体推迟了实现，并指定原因。*/
 @TodoMarker
+@JvmSynthetic
 inline fun DELAY(reason: String) = DELAY(reason) { Unit }
 
 /**表明一个操作推迟了实现。返回模拟结果。*/
 @TodoMarker
+@JvmSynthetic
 inline fun <T> DELAY(lazyDummyResult: () -> T): T = lazyDummyResult().also {
 	println("An operation is delay implemented.".let { "\u001B[33m$it\u001B[0m" })
 	println("Location: $currentClassFullName".let { "\u001B[33m$it\u001B[0m" })
@@ -24,6 +27,7 @@ inline fun <T> DELAY(lazyDummyResult: () -> T): T = lazyDummyResult().also {
 
 /**表明一个方法体推迟了实现，并指定原因。返回模拟结果。*/
 @TodoMarker
+@JvmSynthetic
 inline fun <T> DELAY(reason: String, lazyDummyResult: () -> T): T = lazyDummyResult().also {
 	println("An operation is delay implemented: $reason".let { "\u001B[33m$it\u001B[0m" })
 	println("Location: $currentClassFullName".let { "\u001B[33m$it\u001B[0m" })
@@ -31,6 +35,7 @@ inline fun <T> DELAY(reason: String, lazyDummyResult: () -> T): T = lazyDummyRes
 
 /**表明一个方法体中存在问题。*/
 @TodoMarker
+@JvmSynthetic
 inline fun FIXME() = run {
 	println("An operation has an unresolved issue.".let { "\u001B[91m$it\u001B[0m" })
 	println("Location: $currentClassFullName".let { "\u001B[91m$it\u001B[0m" })
@@ -38,6 +43,7 @@ inline fun FIXME() = run {
 
 /**表明一个方法体中存在问题，并指明原因。*/
 @TodoMarker
+@JvmSynthetic
 inline fun FIXME(message: String) = run {
 	println("An operation has an unresolved issue: $message".let { "\u001B[91m$it\u001B[0m" })
 	println("Location: $currentClassFullName".let { "\u001B[91m$it\u001B[0m" })
@@ -48,6 +54,7 @@ inline fun FIXME(message: String) = run {
 
 //region Standard.kt extensions (Scope functions)
 /**当满足条件时，以接收者为代码体的接收者，执行一段代码并返回转化后的结果，否则返回自身。*/
+@JvmSynthetic
 inline fun <T> T.run(condition: Boolean, block: T.() -> T): T {
 	contract {
 		callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -56,6 +63,7 @@ inline fun <T> T.run(condition: Boolean, block: T.() -> T): T {
 }
 
 /**当满足条件时，以接收者为代码体的参数，执行一段代码并返回转化后的结果，否则返回自身。*/
+@JvmSynthetic
 inline fun <T> T.let(condition: Boolean, block: (T) -> T): T {
 	contract {
 		callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -64,6 +72,7 @@ inline fun <T> T.let(condition: Boolean, block: (T) -> T): T {
 }
 
 /**尝试执行一段代码，并在发生异常时打印堆栈信息。*/
+@JvmSynthetic
 inline fun tryOrPrint(block: () -> Unit) {
 	contract {
 		callsInPlace(block, InvocationKind.EXACTLY_ONCE)
@@ -76,6 +85,7 @@ inline fun tryOrPrint(block: () -> Unit) {
 }
 
 /**尝试执行一段代码，并忽略异常。*/
+@JvmSynthetic
 inline fun tryOrIgnore(block: () -> Unit) {
 	contract {
 		callsInPlace(block, InvocationKind.EXACTLY_ONCE)
@@ -89,6 +99,7 @@ inline fun tryOrIgnore(block: () -> Unit) {
 @PublishedApi internal var enableOnce = false
 
 /**执行一段代码且仅执行一次。可指定是否重置单次状态。*/
+@JvmSynthetic
 inline fun once(resetStatus: Boolean = false, block: () -> Unit) {
 	contract {
 		callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -102,6 +113,7 @@ inline fun once(resetStatus: Boolean = false, block: () -> Unit) {
 
 //region Precondition.kt extensions
 /**如果判定失败，则抛出一个[UnsupportedOperationException]。*/
+@JvmSynthetic
 inline fun accept(value: Boolean) {
 	contract {
 		returns() implies value
@@ -110,6 +122,7 @@ inline fun accept(value: Boolean) {
 }
 
 /**如果判定失败，则抛出一个[UnsupportedOperationException]，带有懒加载的信息。*/
+@JvmSynthetic
 inline fun accept(value: Boolean, lazyMessage: () -> Any) {
 	contract {
 		returns() implies value
@@ -121,6 +134,7 @@ inline fun accept(value: Boolean, lazyMessage: () -> Any) {
 }
 
 /**如果判定失败，则抛出一个[UnsupportedOperationException]。*/
+@JvmSynthetic
 inline fun <T> acceptNotNull(value: T?) {
 	contract {
 		returns() implies (value != null)
@@ -129,6 +143,7 @@ inline fun <T> acceptNotNull(value: T?) {
 }
 
 /**如果判定失败，则抛出一个[UnsupportedOperationException]，带有懒加载的信息。*/
+@JvmSynthetic
 inline fun <T> acceptNotNull(value: T?, lazyMessage: () -> Any): T {
 	contract {
 		returns() implies (value != null)