+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/SystemExtensions.kt
@@ -14,6 +14,7 @@ internal fun getEngine(language: String): ScriptEngine {
 	val fixedLanguage = language.toLowerCase()
 	return scriptEngines.getOrPut(fixedLanguage) {
 		engineManager.getEngineByName(fixedLanguage) ?: engineManager.getEngineByExtension(fixedLanguage)
+		?: throw UnsupportedOperationException("No script engine library found for target script language.")
 	}
 }
 
@@ -34,6 +35,7 @@ inline fun <reified T> eval(language: String, context: ScriptContext, lazyScript
 
 
 /**执行一段懒加载的命令。可指定工作目录，默认为当前目录；可指定环境变量，默认为空。*/
+@JvmSynthetic
 inline fun exac(vararg environmentVariables: String, workDirectory: File? = null, lazyCommand: () -> String): Process {
 	return Runtime.getRuntime().exec(lazyCommand(), environmentVariables, workDirectory)
 }