+++ breeze-generator/src/main/kotlin/com/windea/breezeframework/generator/code/SqlGenerator.kt
@@ -2,6 +2,7 @@
 
 package com.windea.breezeframework.generator.code
 
+import com.windea.breezeframework.core.enums.core.*
 import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.data.enums.*
 import com.windea.breezeframework.data.serializers.*
@@ -19,7 +20,7 @@ object SqlGenerator : Generator {
 		val inputMap = inputType.serializer.load<SqlDataMap>(inputText)
 		return getSqlDataString(inputMap)
 	}
-	
+
 	/**
 	 * 根据输入文件和输入数据类型，生成Sql数据到指定输出文件。默认使用Yaml类型。
 	 *
@@ -29,24 +30,24 @@ object SqlGenerator : Generator {
 		val inputMap = inputType.serializer.load<SqlDataMap>(inputFile)
 		outputFile.writeText(getSqlDataString(inputMap))
 	}
-	
+
 	private fun getSqlDataString(inputMap: SqlDataMap): String {
 		val databaseName = inputMap.keys.first()
 		val database = inputMap.values.first()
-		
+
 		return """
 		use $databaseName;
-		
+
 		${database.joinToString("\n\n") { (tableName, table) ->
 			val columnNamesSnippet = table.first().keys.joinToString()
-			
+
 			"""
 			insert into $tableName ($columnNamesSnippet) values
 			${table.joinToString(",\n", "", ";\n") { data ->
 				val columnsSnippet = data.values.joinToString {
-					it.toString().wrapQuote('\'').escapeBy()
+					it.toString().wrapQuote('\'').escapeBy(EscapeType.Java)
 				}
-				
+
 				"""  ($columnsSnippet)"""
 			}}
 			""".trimRelativeIndent()