+++ breeze-generator/src/main/kotlin/com/windea/breezeframework/generator/specific/IdeaConfigGenerator.kt
@@ -2,6 +2,7 @@
 
 package com.windea.breezeframework.generator.specific
 
+import com.windea.breezeframework.core.enums.core.*
 import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.data.enums.*
 import com.windea.breezeframework.data.serializers.*
@@ -19,7 +20,7 @@ object IdeaConfigGenerator : Generator {
 		val inputMap = inputType.serializer.load<SchemaDefinitionMap>(inputText)
 		return getYamlAnnotationString(inputMap)
 	}
-	
+
 	/**
 	 * 根据输入文件和输入数据类型，生成自定义Yaml注解的动态模版配置文件文本。默认使用Yaml类型。
 	 *
@@ -29,23 +30,23 @@ object IdeaConfigGenerator : Generator {
 		val inputMap = inputType.serializer.load<SchemaDefinitionMap>(inputFile)
 		outputFile.writeText(getYamlAnnotationString(inputMap))
 	}
-	
+
 	private fun getYamlAnnotationString(inputMap: SchemaDefinitionMap): String {
 		val definitions = inputMap["definitions"] as SchemaMap
 		return """
 		<templateSet group="YamlAnnotation">
 		${definitions.joinToString("\n\n") { (templateName, template) ->
-			val description = (template.getOrDefault("description", "") as String).escapeBy()
+			val description = (template.getOrDefault("description", "") as String).escapeBy(EscapeType.Java)
 			val params = if("properties" in template) template["properties"] as Map<String, Map<String, Any?>> else mapOf()
 			val paramSnippet = if(params.isEmpty()) "" else ": {${params.keys.joinToString(", ") { "$it: $$it$" }}}"
-			
+
 			"""
 			  <template name="@$templateName" value="@$templateName$paramSnippet"
 		                description="$description"
 		                toReformat="true" toShortenFQNames="true" useStaticImport="true">${
 			params.joinToString("\n") { (paramName, param) ->
-				val defaultValue = (param.getOrDefault("default", "") as String).escapeBy()
-				
+				val defaultValue = (param.getOrDefault("default", "") as String).escapeBy(EscapeType.Java)
+
 				"""    <variable name="$paramName" expression="" defaultValue="&quot;$defaultValue&quot;" alwaysStopAt="true"/>"""
 			}.ifNotEmpty { "\n$it" }}
 			    <context>