+++ breeze-spring-boot/src/main/kotlin/com/windea/breezeframework/springboot/mail/MailExtensions.kt
@@ -5,6 +5,7 @@ package com.windea.breezeframework.springboot.mail
 import org.springframework.mail.javamail.*
 
 /**发送邮件。基于[MimeMessageHelper]。*/
+@JvmOverloads
 fun JavaMailSender.sendEmail(encoding: String? = null, prepare: MimeMessageHelper.() -> Unit) {
 	try {
 		this.send { MimeMessageHelper(it, encoding).prepare() }
@@ -14,6 +15,7 @@ fun JavaMailSender.sendEmail(encoding: String? = null, prepare: MimeMessageHelpe
 }
 
 /**发送邮件。基于[MimeMessageHelper]。*/
+@JvmOverloads
 fun JavaMailSender.sendEmail(multipart: Boolean, encoding: String? = null, prepare: MimeMessageHelper.() -> Unit) {
 	try {
 		this.send { MimeMessageHelper(it, multipart, encoding).prepare() }
@@ -23,6 +25,7 @@ fun JavaMailSender.sendEmail(multipart: Boolean, encoding: String? = null, prepa
 }
 
 /**发送邮件。基于[MimeMessageHelper]。*/
+@JvmOverloads
 fun JavaMailSender.sendEmail(multipartMode: Int, encoding: String? = null, prepare: MimeMessageHelper.() -> Unit) {
 	try {
 		this.send { MimeMessageHelper(it, multipartMode, encoding).prepare() }