+++ breeze-text/src/main/kotlin/com/windea/breezeframework/text/extensions/NumberAndAbbreviationExtensions.kt
@@ -9,18 +9,21 @@ import kotlin.math.*
 
 /**转化为简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Int.toAbbreviation(scale: Int, precision: Int = 0): String {
 	return this.toFloat().toAbbreviation(scale, precision)
 }
 
 /**转化为简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Long.toAbbreviation(scale: Int, precision: Int = 0): String {
 	return this.toFloat().toAbbreviation(scale, precision)
 }
 
 /**转化为简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Float.toAbbreviation(scale: Int, precision: Int = 0): String {
 	require(scale in 1..9) { "Scale must between 1 and 9, but was $scale." }
 
@@ -31,6 +34,7 @@ fun Float.toAbbreviation(scale: Int, precision: Int = 0): String {
 
 /**转化为简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Double.toAbbreviation(scale: Int, precision: Int = 0): String {
 	return this.toFloat().toAbbreviation(scale, precision)
 }
@@ -54,18 +58,21 @@ private fun getAbbreviation(valueSnippet: String, scale: Int): String {
 
 /**转化为中文简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Int.toChsAbbreviation(scale: Int, precision: Int = 0): String {
 	return this.toFloat().toChsAbbreviation(scale, precision)
 }
 
 /**转化为中文简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Long.toChsAbbreviation(scale: Int, precision: Int = 0): String {
 	return this.toFloat().toChsAbbreviation(scale, precision)
 }
 
 /**转化为中文简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Float.toChsAbbreviation(scale: Int, precision: Int = 0): String {
 	require(scale in 1..9) { "Scale must between 1 and 9, but was $scale." }
 
@@ -76,6 +83,7 @@ fun Float.toChsAbbreviation(scale: Int, precision: Int = 0): String {
 
 /**转化为中文简写形式。可指定1到9的缩放级别，可指定默认为0的精确度。*/
 @NotRecommended("Use java12's NumberFormat.getCompactNumberInstance() to format numbers.")
+@JvmOverloads
 fun Double.toChsAbbreviation(scale: Int, precision: Int = 0): String {
 	return this.toFloat().toChsAbbreviation(scale, precision)
 }